Adiciona persistencia no MySQL

+++ src/main/java/br/com/elo7/sonda/candidato/restapi/controller/dto/PlanetDTO.java
@@ -2,17 +2,17 @@ package br.com.elo7.sonda.candidato.restapi.controller.dto;
 
 public class PlanetDTO {
 
-    private int id;
+    private long id;
     private int width;
     private int height;
 
-    public PlanetDTO(int id, int width, int height) {
+    public PlanetDTO(long id, int width, int height) {
         this.id = id;
         this.width = width;
         this.height = height;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
